+++ examples/frame_metalayers.c
@@ -41,11 +41,11 @@
 int main(void) {
   blosc_init();
 
-  size_t isize = CHUNKSIZE * sizeof(int32_t);
+  int32_t isize = CHUNKSIZE * sizeof(int32_t);
   int32_t* data = malloc(isize);
   int64_t nbytes, cbytes;
   int i, nchunk;
-  int nchunks;
+  int64_t nchunks;
   blosc_timestamp_t last, current;
   double ttotal;
 
@@ -83,9 +83,9 @@ int main(void) {
   blosc_set_timestamp(&current);
   ttotal = blosc_elapsed_secs(last, current);
   printf("Compression ratio: %.2f MB -> %.2f MB (%.1fx)\n",
-         nbytes / MB, cbytes / MB, (1. * nbytes) / cbytes);
+         (double)nbytes / MB, (double)cbytes / MB, (1. * (double)nbytes) / (double)cbytes);
   printf("Compression time: %.3g s, %.1f MB/s\n",
-         ttotal, nbytes / (ttotal * MB));
+         ttotal, (double)nbytes / (ttotal * MB));
 
   blosc_set_timestamp(&last);
 
@@ -104,7 +104,7 @@ int main(void) {
   blosc_set_timestamp(&current);
   ttotal = blosc_elapsed_secs(last, current);
   printf("Time for frame -> fileframe (simple_frame.b2frame): %.3g s, %.1f GB/s\n",
-         ttotal, nbytes / (ttotal * GB));
+         ttotal, (double)nbytes / (ttotal * GB));
 
   // fileframe (file) -> schunk2 (schunk based on a on-disk frame)
   blosc_set_timestamp(&last);
@@ -116,7 +116,7 @@ int main(void) {
   blosc_set_timestamp(&current);
   ttotal = blosc_elapsed_secs(last, current);
   printf("Time for fileframe (%s) -> schunk : %.3g s, %.1f GB/s\n",
-         schunk2->storage->urlpath, ttotal, nbytes / (ttotal * GB));
+         schunk2->storage->urlpath, ttotal, (double)nbytes / (ttotal * GB));
 
   // Check that the metalayers had a good roundtrip
   if (schunk2->nmetalayers != 2) {
