+++ plugins/codecs/ndlz/ndlz4x4.c
@@ -690,5 +690,5 @@ int ndlz4_decompress(const uint8_t *input, int32_t input_len, uint8_t *output, i
     return 0;
   }
 
-  return ind;
+  return (int)ind;
 }
